requirements.txt
Flask
Flask_SQLAlchemy
psycopg2-binary
python-dotenv
flask-bcrypt
flask-jwt-extended
flask-cors
bcrypt
jwt
